added new voices
new voices were released for the model aswell as an update to the model itself. The new voice files offer different language options but my phenomizer will need to be reworked. moreover, the new model v1 has come out i plan to l to test, convert and further quantize. additionally i added a script for converting the voices from pt to json with python and torch.